underfoot init

+++ MQ2EQIM/ISXEQEQIM.h
@@ -3,27 +3,27 @@
 
 
 class ISXEQEQIM :
-	public ISXInterface
+    public ISXInterface
 {
 public:
 
-	virtual bool Initialize(ISInterface *p_ISInterface);
-	virtual void Shutdown();
+    virtual bool Initialize(ISInterface *p_ISInterface);
+    virtual void Shutdown();
 
-	void LoadSettings();
-	void ConnectServices();
-	void RegisterCommands();
-	void RegisterAliases();
-	void RegisterDataTypes();
-	void RegisterTopLevelObjects();
-	void RegisterServices();
+    void LoadSettings();
+    void ConnectServices();
+    void RegisterCommands();
+    void RegisterAliases();
+    void RegisterDataTypes();
+    void RegisterTopLevelObjects();
+    void RegisterServices();
 
-	void DisconnectServices();
-	void UnRegisterCommands();
-	void UnRegisterAliases();
-	void UnRegisterDataTypes();
-	void UnRegisterTopLevelObjects();
-	void UnRegisterServices();
+    void DisconnectServices();
+    void UnRegisterCommands();
+    void UnRegisterAliases();
+    void UnRegisterDataTypes();
+    void UnRegisterTopLevelObjects();
+    void UnRegisterServices();
 
 };
 
@@ -41,11 +41,11 @@ extern HISXSERVICE hEQZoneService;
 extern ISXEQEQIM *pExtension;
 #define printf pISInterface->Printf
 
-#define EzDetour(Address, Detour, Trampoline) IS_Detour(pExtension,pISInterface,hMemoryService,(unsigned long)Address,Detour,Trampoline)
-#define EzUnDetour(Address) IS_UnDetour(pExtension,pISInterface,hMemoryService,(unsigned long)Address)
+#define EzDetour(Address, Detour, Trampoline) IS_Detour(pExtension,pISInterface,hMemoryService,(unsigned int)Address,Detour,Trampoline)
+#define EzUnDetour(Address) IS_UnDetour(pExtension,pISInterface,hMemoryService,(unsigned int)Address)
 
-#define EzModify(Address,NewData,Length,Reverse) Memory_Modify(pExtension,pISInterface,hMemoryService,(unsigned long)Address,NewData,Length,Reverse)
-#define EzUnModify(Address) Memory_UnModify(pExtension,pISInterface,hMemoryService,(unsigned long)Address)
+#define EzModify(Address,NewData,Length,Reverse) Memory_Modify(pExtension,pISInterface,hMemoryService,(unsigned int)Address,NewData,Length,Reverse)
+#define EzUnModify(Address) Memory_UnModify(pExtension,pISInterface,hMemoryService,(unsigned int)Address)
 
 #define EzHttpRequest(_URL_,_pData_) IS_HttpRequest(pExtension,pISInterface,hHTTPService,_URL_,_pData_)
